Successful login checkpoint.
There was something wrong with the auth.ts in the frontend. This update
solved it and at least got to the point where I could login using the
Supabase JWTs and the barista profiles. There's an open issue where if I
refresh a page, it doesn't successfully maintain my logged in status.
I'll try to fix that next commit.